backend/requirements.txt
Django==5.0.4
djangorestframework==3.15.1
psycopg2-binary==2.9.9
gunicorn==22.0.0
requests==2.32.3     # for Groq API calls
